split running of gba and added nds section to bin
